+++ pysparkling/rdd.py
@@ -834,6 +834,86 @@ class RDD(object):
         TextFile(path+'/_SUCCESS').dump()
         return self
 
+    def sortBy(self, keyfunc, ascending=True, numPartitions=None):
+        """
+        :param keyfunc:
+            Returns the value that will be sorted.
+
+        :param ascending:
+            Default is True.
+
+        :param numPartitions:
+            Default is None. None means the output will have the same number of
+            partitions as the input.
+
+        :returns:
+            A new sorted RDD.
+
+        .. note::
+            Sorting is currently implemented as a local operation.
+
+
+        Examples:
+
+        >>> from pysparkling import Context
+        >>> rdd = Context().parallelize([5, 1, 2, 3])
+        >>> rdd.sortBy(lambda x: x).collect()[0]
+        1
+
+        >>> from pysparkling import Context
+        >>> rdd = Context().parallelize([1, 5, 2, 3])
+        >>> rdd.sortBy(lambda x: x, ascending=False).collect()[0]
+        5
+
+        """
+
+        if numPartitions is None:
+            numPartitions = self.getNumPartitions()
+
+        return self.context.parallelize(
+            sorted(self.collect(), key=keyfunc, reverse=not ascending),
+            numPartitions,
+        )
+
+    def sortByKey(self, ascending=True, numPartitions=None,
+                  keyfunc=lambda x: x[0]):
+        """
+        :param ascending:
+            Default is True.
+
+        :param numPartitions:
+            Default is None. None means the output will have the same number of
+            partitions as the input.
+
+        :param keyfunc:
+            Returns the value that will be sorted.
+
+        :returns:
+            A new sorted RDD.
+
+        .. note::
+            Sorting is currently implemented as a local operation.
+
+
+        Examples:
+
+        >>> from pysparkling import Context
+        >>> rdd = Context().parallelize(
+        ...     [(5, 'a'), (1, 'b'), (2, 'c'), (3, 'd')]
+        ... )
+        >>> rdd.sortByKey().collect()[0][1] == 'b'
+        True
+
+        >>> from pysparkling import Context
+        >>> rdd = Context().parallelize(
+        ...     [(1, 'b'), (5, 'a'), (2, 'c'), (3, 'd')]
+        ... )
+        >>> rdd.sortByKey(ascending=False).collect()[0][1] == 'a'
+        True
+
+        """
+        return self.sortBy(keyfunc, ascending, numPartitions)
+
     def stats(self):
         """
         :returns: